added time analysis code

+++ mainFile2.py
@@ -3,8 +3,11 @@ import os
 from subprocess import Popen
 import zipfile
 
-from task4 import *
+from timeplots import *
+
+
 gen_directory = "/Users/udaisingh/Desktop/gen"
+download = True
 
 FIPS = ["02"]
 
@@ -19,33 +22,35 @@ FIPS = ["01","02", "04", "05", "06", "08",
 """
 
 def process(fip):
-    print("Current FIP processed: " + fip)
-    download_command = "aws s3 cp s3://zillow-data-raw/20191001/" + fip + ".zip ."
-    #download zip file
-    os.system(download_command)
-    print("Zip File Downloaded")
-    
-    #unzip file
-    fileName = fip + ".zip"
-    unzip_command = "unzip " + fileName + " -d " + fileName.strip(".zip") 
-    os.system(unzip_command)
-    print("File Unzipped")
-    
-    folder = fileName.strip(".zip") + "/"
-    copy_command = "cp Layout.xlsx " + str(folder)
-    os.system(copy_command)
-    
-    #remove zip file
-    remove_zip_file = "rm -rf " + fileName
-    os.system(remove_zip_file)
+    if (download):
+        print("Current FIP processed: " + fip)
+        download_command = "aws s3 cp s3://zillow-data-raw/20191001/" + fip + ".zip ."
+        #download zip file
+        os.system(download_command)
+        print("Zip File Downloaded")
+        
+        #unzip file
+        fileName = fip + ".zip"
+        unzip_command = "unzip " + fileName + " -d " + fileName.strip(".zip") 
+        os.system(unzip_command)
+        print("File Unzipped")
+        
+        folder = fileName.strip(".zip") + "/"
+        copy_command = "cp Layout.xlsx " + str(folder)
+        os.system(copy_command)
+        
+        #remove zip file
+        remove_zip_file = "rm -rf " + fileName
+        os.system(remove_zip_file)
 
     print("Program Begining")
     gen_dir = os.path.join(gen_directory, fip)
     program(fip, gen_dir)
     print("Program Ended")
 
-    remove_folder_command = "rm -rf " + fip + "/"
-    os.system(remove_folder_command)    
+    if (download):
+        remove_folder_command = "rm -rf " + fip + "/"
+        os.system(remove_folder_command)    
 
 
 for fip in FIPS:
